Connect wallets from get address button on Bridge Page (#789)
* wip

* use connect wallet button > icon
